Add Python script to update mod.json
