+++ resources/update-words.py
@@ -3,52 +3,37 @@
 from argparse import ArgumentParser, FileType
 from datetime import datetime, timezone
 from requests import get
-from bs4 import BeautifulSoup
+from os import path
 
 import json
 import re
 
 
-def update_words(output):
+def update_words(mod_json):
 	"""Updates the word lists"""
 
-	# First get the HTML and figure out what the current JS is
-	html = get("https://www.nytimes.com/games/wordle/index.html").text
-	soup = BeautifulSoup(html, "html.parser")
-	script, = [script["src"] for script in soup.findAll("script") if script.has_attr("src") and script["src"].startswith("https://www.nytimes.com/games-assets/v2/wordle.")]
+	j = {}
 
-	# So we can get that and grab the base list incase they ever update that again
-	js = get(script).text
-	guesses = json.loads(re.findall(r'\["aahed"(?:,"\w{5}")*(?=,"cigar")', js)[0] + "]")
-	guesses.sort()
-	choices = json.loads("[" + re.findall(r'"cigar"(?:,"\w{5}")*\]', js)[0])
+	# Open the mod.json if it exists
+	if mod_json and path.exists(mod_json):
+		with open(mod_json, "r") as f:
+			j = json.load(f)
 
-	# Then grab the known word order
-	words = get("https://wordle.xn--rck9c.xn--tckwe/words.php?date=2021-06-19&limit=10000").json()
+	# Ensure the structure exists
+	if "words" not in j:
+			j["words"] = {}
 
-	# Now let's make the file, first inclues
-	output.write('#include "words.hpp"\n\n')
+	# Get all of the known word order
+	j["words"]["order"] = get("https://wordle.xn--rck9c.xn--tckwe/words.php?date=2021-06-19&limit=10000&include=id").json()
 
-	# Then the known word order
-	output.write("std::vector<int> Words::order = {")
-	output.write(", ".join([word["id"] for word in words]))
-	output.write("};\n\n")
-
-	# The choice words
-	output.write('std::vector<std::u16string> Words::choices = {u"')
-	output.write('", u"'.join([word.upper() for word in choices]))
-	output.write('"};\n\n')
-
-	# And the guess words
-	output.write('std::vector<std::u16string> Words::guesses = {u"')
-	output.write('", u"'.join([word.upper() for word in guesses]))
-	output.write('"};\n')
+	with open(mod_json, "w") as f:
+		json.dump(j, f)
 
 
 if __name__ == "__main__":
 	parser = ArgumentParser(description="Updates the Wordle DS word list")
-	parser.add_argument("output", metavar="output.cpp", type=FileType("w"), help="output C++ source")
+	parser.add_argument("--file", "-f", default="mod.json", help="mod.json file to update")
 
 	args = parser.parse_args()
 
-	update_words(args.output)
+	update_words(args.file)
